restructued CheckConjugation component, added in SimpleIndicative/SimpleSubjunc component, and organized results into a grid
