+++ src/App.css
@@ -102,7 +102,6 @@
 
 /* CheckConjugation */
 .conjugationsMainCon {
-	border: 2px solid red;
 	width: 95%;
 	margin: auto;
 	height: auto;
@@ -113,11 +112,16 @@
 	display: flex;
 }
 
-.conjugations {
-	border: 2px solid red;
+.conjugationsView {
 	display: flex;
 	flex-direction: column;
 	width: 20%;
+}
+
+.conjugations {
+	display: flex;
+	flex-direction: column;
+	width: 80%;
 }
 /* .conjugations .title {
 	display: flex;
